Allow already replicated records to be removed from pending_changes
Remove savepoint setup and teardown.  We're already in a transaction block, this is causing a lot of errors when the connection is reset from under it.
Rescue ActiveRecord::RecordNotUnique errors allowing already replicated records to be removed from the pending_changes table.
Remove old difference loader code that is no longer needed, but still querying the database.

https://bugzilla.redhat.com/show_bug.cgi?id=1119453